clean docs
