--- docs/src/css/appearance-option.css
@@ -1,22 +1,0 @@
-.AppearanceOption {
-  background-color: white;
-  box-shadow: inset 0 0 0 1px var(--color-border-default);
-  padding: 16px;
-  margin-bottom: 16px;
-}
-
-.AppearanceOption-component {
-  border-bottom: 1px solid var(--color-border-muted);
-  padding-bottom: 16px;
-}
-
-.AppearanceOption-description {
-  & h3 {
-    margin-top: 16px;
-    margin-bottom: 0;
-
-    & + p {
-      margin-top: 8px;
-    }
-  }
-}
