--- docs/src/css/component-readme.css
@@ -1,70 +1,0 @@
-.ComponentReadme {
-}
-
-.ComponentReadme-inner {
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 40px;
-}
-
-.ComponentReadme-title {
-  font-size: 32px;
-  font-weight: 400;
-  line-height: 32px;
-  letter-spacing: -0.2px;
-  margin-bottom: 4px;
-}
-
-.ComponentReadme-subtitle {
-  font-size: 16px;
-  font-weight: 300;
-  line-height: 28px;
-}
-
-.ComponentReadme-header {
-  border-bottom: 1px solid var(--neutral-20A);
-  padding-bottom: 16px;
-  margin-bottom: 24px;
-}
-
-.ComponentReadme-githubLink {
-  color: var(--blue-500);
-
-  &:hover {
-    text-decoration: underline;
-  }
-
-  &:focus {
-    outline: 1px solid var(--blue-500);
-  }
-}
-
-/**
- * Not used atm
- */
-.ComponentReadme-nav {
-  margin-top: 16px;
-  border-bottom: 1px solid var(--neutral-20A);
-
-  & a {
-    color: var(--neutral-300);
-    font-size: 14px;
-    display: inline-block;
-    padding: 12px 0 12px 0;
-    margin-right: 24px;
-
-    &:hover {
-      color: var(--neutral-800);
-    }
-
-    &[aria-selected='true'] {
-      color: var(--neutral-800);
-      box-shadow: 0 1px var(--neutral-100);
-    }
-  }
-}
-
-.ComponentSection {
-  margin-top: 64px;
-  margin-bottom: 64px;
-}
