+++ docs/src/css/mdx-page.css
@@ -1,5 +1,4 @@
 .MDXPage-header {
-  /* background-color: var(--N1); */
   border-bottom: var(--color-border-default);
 }
