+++ docs/src/pages/get-started/introduction.js
@@ -3,7 +3,7 @@ import React from 'react'
 import Helmet from 'react-helmet'
 import TopBar from '../../components/TopBar'
 import GetStartedSidebar from '../../components/GetStartedSidebar'
-import SyntaxHighlighter from '../../components/SyntaxHighlighter2'
+import SyntaxHighlighter from '../../components/SyntaxHighlighter'
 import Layout from '../../components/Layout'
 
 const NativeLink = ({ ...props }) => {
